requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
DAWG-Python==0.7.2
docopt==0.6.2
Faker==21.0.0
fastapi==0.105.0
fonttools==4.47.0
h11==0.14.0
idna==3.6
kiwisolver==1.4.5
lxml==4.9.4
matplotlib==3.8.2
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
pdf2image==1.16.3
Pillow==10.1.0
pydantic==2.5.3
pydantic_core==2.14.6
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pymorphy3==1.3.1
pymorphy3-dicts-ru==2.4.417150.4580142
pyparsing==3.1.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.25.0
